app/requirement.txt
streamlit
pandas
joblib
tensorflow
google-generativeai
